Replace table lists with string buffers
This replaces much of the code which generates the trac output with
internal string buffers rather than lists.  It also modernises the
code in various places, such as using instance attributes in place of
calling arguments etc.

This is an interim commit.  There are a number of items that still
need to be fixed or modernised, and exception generation and handling
leaves a lot to be desired.