Bump the error handling up to the top level
Move the exception handling from print_report up to the top level main
function.  This makes it possible to log a full trace of the stack from
top to bottom.

It also adds exception handling for other bits of suite report,
ensuring that something ends up in the log in almost every case.