Make function names non-private
There is no reason for functions visible to the entirety of
suite_report to be private, so remove leading underscores.  This also
makes test functions look a bit cleaner.

Change the name of an optional argument in the _unapck_suite_value
method call from remove_quotes - now the name of a function - to
dequote.